requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
iniconfig==2.3.0
kombu==5.6.2
Mako==1.3.11
MarkupSafe==3.0.3
packaging==26.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uvicorn==0.45.0
vine==5.1.0
wcwidth==0.6.0
